Fix version tracking issue for a ui file
